pyproject.toml
[project]
name = "customers-balance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-cotton>=2.1.3",
    "django==5.1.0",
    "django-extensions>=4.1",
    "django-vite>=3.1.0",
    "django-filter>=25.1",
    "django-import-export[xlsx]>=4.3.10",
    "django-htmx>=1.26.0",
    "mysqlclient>=2.2.7",
    "django-widget-tweaks>=1.5.0",
    "python-decouple>=3.8",
    "django-solo>=2.4.0",
    "cotton-icons>=0.1.0",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "django-silk>=5.4.3",
    "parameterized>=0.9.0",
]
